Move abstract_tensor
