+++ tests/cpp/test_gpu_view.cpp
@@ -9,7 +9,6 @@
 #include <gmock/gmock-matchers.h>
 #include <gtest/gtest.h>
 
-#include <abstract_tensor.h>
 #include <codegen.h>
 #include <device_lower/analysis/divisible_split.h>
 #include <device_lower/lower2device.h>
@@ -32,6 +31,7 @@
 #include <runtime/fusion_executor_cache.h>
 #include <scheduler/all_schedulers.h>
 #include <scheduler/reduction_utils.h>
+#include <scheduler/tools/abstract_tensor.h>
 #include <scheduler/tools/inlining.h>
 #include <scheduler/utils.h>
 #include <tests/cpp/utils.h>
