+++ tests/cpp/test_indexing.cpp
@@ -12,7 +12,6 @@
 #include <tests/cpp/utils.h>
 #include <tests/cpp/validator.h>
 
-#include <abstract_tensor.h>
 #include <fusion.h>
 #include <id_model/id_model.h>
 #include <id_model/indexing.h>
@@ -22,6 +21,7 @@
 #include <ir/builder.h>
 #include <kernel_ir_dispatch.h>
 #include <ops/all_ops.h>
+#include <scheduler/tools/abstract_tensor.h>
 #include <scheduler/tools/inlining.h>
 #include <scheduler/utils.h>
