+++ tests/cpp/test_swizzle.cpp
@@ -8,10 +8,10 @@
 #include <csrc/exceptions.h>
 #include <gtest/gtest.h>
 
-#include <abstract_tensor.h>
 #include <device_lower/analysis/bank_conflict.h>
 #include <ops/all_ops.h>
 #include <runtime/fusion_executor_cache.h>
+#include <scheduler/tools/abstract_tensor.h>
 #include <scheduler/tools/inlining.h>
 #include <swizzle.h>
 #include <tests/cpp/utils.h>
