+++ tests/cpp/test_utils.cpp
@@ -9,11 +9,11 @@
 #include <gmock/gmock-matchers.h>
 #include <gtest/gtest.h>
 
-#include <abstract_tensor.h>
 #include <device_lower/utils.h>
 #include <fusion.h>
 #include <ops/all_ops.h>
 #include <runtime/executor_utils.h>
+#include <scheduler/tools/abstract_tensor.h>
 #include <scheduler/utils.h>
 #include <scheduler/vectorize_helper.h>
 #include <tests/cpp/utils.h>
